chore: optimize integration tests
