+++ src/main/resources/application-queuedtesting.properties
@@ -1,2 +1,2 @@
 live.event.retry.send.cron=* * * * * ?
-live.event.retry.amount=0
+live.event.retry.amount=5
